+++ src/hh.js
@@ -1,8 +1,7 @@
 const { h } = require('snabbdom');
 
 const hh = (name, values, children) => {
-    console.log('hh',{name, values, children})
-    return h(name, values, children)
+    return h(name, values, children);
 };
 
 module.exports = hh;