+++ src/node-components/def-scene.js
@@ -1,9 +1,20 @@
 const scene = {
     tagName: 'scene',
-    props: {},
+    props: {
+        clearColor: {
+            setter: (oldValue, newValue, node, cmp, { BABYLON }) =>
+                (cmp.clearColor = BABYLON.Vector3(...newValue)),
+        },
+    },
 
-    creator: (propsSetters, { BABYLON, canvas, engine, scene }, node) => () => {
+    creator: (propsSetters, context, node) => () => {
+        const { scene } = context;
         const cmp = scene;
+
+        if (node.clearColor) {
+            propsSetters.clearColor.setter(node.clearColor, node.clearColor, node, cmp, context);
+        }
+
         return cmp;
     },
 };