+++ test/component-create.spec.js
@@ -1,20 +1,20 @@
 describe('COMPONENT', function() {
-    const { createComponent } = require('../src/component');
+    const { createApp } = require('../src/component');
 
     it('should not initialize without api', () => {
         const expectedError =
             'Invalid paramaters, you need to provide an object with a key named api:\n{ api: createBabylonSceneAPI(...) }';
 
-        expect(createComponent).to.throw(expectedError);
-        expect(createComponent, {}).to.throw(expectedError);
+        expect(createApp).to.throw(expectedError);
+        expect(createApp, {}).to.throw(expectedError);
     });
 
-    it('should create a component', () => {
-        const component = createComponent({
+    it('should create a app', () => {
+        const app = createApp({
             api: {},
         });
 
-        expect(component)
+        expect(app)
             .to.an('object')
             .to.have.all.keys(['render', 'dispose']);
     });