+++ CHANGELOG.md
@@ -1,1 +1,29 @@
-# react-babylonjs
+# Changelog
+
+## [0.0.2] 2017-10-09
+### Components 
+- BABYLON.FxaaPostProcess
+- BABYLON.FreeCamera
+- BABYLON.HemisphericLight
+- BABYLON.StandardMaterial
+- BABYLON.ShaderMaterial
+- BABYLON.Mesh
+- BABYLON.Mesh
+- BABYLON.MeshBuilder
+- BABYLON.Mesh
+- BABYLON.Texture
+- BABYLON.CubeTexture
+- BABYLON.Scene
+
+### Helpers
+```javascript
+const { createBabylonSceneAPI, createApp, hhh } = require('react-babylonjs-3d');
+```
+
+- `createApp` helpes to create the whole think, it's just needs the `BABYLON` and `canvas` instances
+- `createBabylonSceneAPI` used to attach the render to an already created(running) engine and scene
+- `hhh` pragma used to enable the JSX syntax, see at the [repo](https://github.com/viniciusrmcarneiro/react-babylonjs-3d-example) the file *.babelrc* and *webpack.config.js*
+
+
+## [0.0.1.alpha*] 2017-10-01
+### just some tests, nothing really useful